+++ empire/server/core/module_models.py
@@ -44,7 +44,7 @@ class EmpireModuleOption(BaseModel):
         return [str(value) for value in v]
 
 
-class EmpireModuleAuthor(BaseModel):
+class EmpireAuthor(BaseModel):
     name: str
     handle: str
     link: str
@@ -59,7 +59,7 @@ class BofModuleOption(BaseModel):
 class EmpireModule(BaseModel):
     id: str
     name: str
-    authors: list[EmpireModuleAuthor] = []
+    authors: list[EmpireAuthor] = []
     description: str = ""
     software: str = ""
     techniques: list[str] = []
@@ -79,23 +79,3 @@ class EmpireModule(BaseModel):
     enabled: bool = True
     advanced: EmpireModuleAdvanced = EmpireModuleAdvanced()
     compiler_yaml: str | None = None
-
-    def matches(self, query: str, parameter: str = "any") -> bool:
-        query = query.lower()
-        match = {
-            "name": query in self.name.lower(),
-            "description": query in self.description.lower(),
-            "comments": any(query in comment.lower() for comment in self.comments),
-            "authors": any(query in author.lower() for author in self.authors),
-        }
-
-        if parameter == "any":
-            return any(match.values())
-
-        return match[parameter]
-
-    @property
-    def info(self) -> dict:
-        desc = self.dict(include={"name", "authors", "description", "comments"})
-        desc["options"] = [option.model_dump() for option in self.options]
-        return desc