+++ empire/server/plugins/csharpserver/csharpserver.py
@@ -6,31 +6,17 @@ import socket
 import subprocess
 
 import empire.server.common.helpers as helpers
-from empire.server.common.plugins import Plugin
+from empire.server.common.plugins import BasePlugin
 from empire.server.core.plugin_service import PluginService
 
 log = logging.getLogger(__name__)
 
 
-class Plugin(Plugin):
+class Plugin(BasePlugin):
     def onLoad(self):
         self.main_menu = None
         self.csharpserver_proc = None
         self.thread = None
-        self.info = {
-            "Name": "csharpserver",
-            "Authors": [
-                {
-                    "Name": "Anthony Rose",
-                    "Handle": "@Cx01N",
-                    "Link": "https://twitter.com/Cx01N_",
-                }
-            ],
-            "Description": ("Empire C# server for agents."),
-            "Software": "",
-            "Techniques": [""],
-            "Comments": [],
-        }
 
         self.options = {
             "status": {
@@ -56,14 +42,14 @@ class Plugin(Plugin):
     def get_commands(self):
         return self.commands
 
-    def register(self, mainMenu):
+    def register(self, main_menu):
         """
-        any modifications to the mainMenu go here - e.g.
+        any modifications to the main_menu go here - e.g.
         registering functions to be run by user commands
         """
-        self.installPath = mainMenu.installPath
-        self.main_menu = mainMenu
-        self.plugin_service: PluginService = mainMenu.pluginsv2
+        self.installPath = main_menu.installPath
+        self.main_menu = main_menu
+        self.plugin_service: PluginService = main_menu.pluginsv2
 
     def do_csharpserver(self, command):
         """