+++ empire/server/plugins/websockify_server/websockify_server.py
@@ -4,33 +4,16 @@ import logging
 import websockify
 
 import empire.server.common.helpers as helpers
-from empire.server.common.plugins import Plugin
+from empire.server.common.plugins import BasePlugin
 from empire.server.core.plugin_service import PluginService
 
 log = logging.getLogger(__name__)
 
 
-class Plugin(Plugin):
+class Plugin(BasePlugin):
     def onLoad(self):
         self.main_menu = None
         self.csharpserver_proc = None
-        self.info = {
-            "Name": "websockify_server",
-            "Authors": [
-                {
-                    "Name": "Anthony Rose",
-                    "Handle": "@Cx01N",
-                    "Link": "https://twitter.com/Cx01N_",
-                }
-            ],
-            "Description": (
-                "Websockify server for TCP proxy/bridge to connect applications. For example: "
-                "run the websockify server to connect the VNC server to noVNC."
-            ),
-            "Software": "",
-            "Techniques": ["T1090"],
-            "Comments": ["https://github.com/novnc/websockify"],
-        }
 
         self.options = {
             "SourceHost": {
@@ -77,14 +60,14 @@ class Plugin(Plugin):
     def get_commands(self):
         return self.commands
 
-    def register(self, mainMenu):
+    def register(self, main_menu):
         """
-        any modifications to the mainMenu go here - e.g.
+        any modifications to the main_menu go here - e.g.
         registering functions to be run by user commands
         """
-        self.installPath = mainMenu.installPath
-        self.main_menu = mainMenu
-        self.plugin_service: PluginService = mainMenu.pluginsv2
+        self.installPath = main_menu.installPath
+        self.main_menu = main_menu
+        self.plugin_service: PluginService = main_menu.pluginsv2
 
     def do_websockify(self, command):
         """